app-feira-livre/requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.2.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djongo==1.2.31
dnspython==2.7.0
PyJWT==2.9.0
pymongo==4.13.0
python-decouple==3.8
sqlparse==0.5.3

